Add load procedure
The load procedure evaluates expressions from files. It always evaluates
them in the global environment, because otherwise functions defined in
files "disappear" after the scope it was loaded in ends.

The function first looks directly for the file named as the first
argument, but if that fails it adds ".scm" to the end of the file name and
tries that file instead.