Add byte type
Necessary for reading from a file in a sane way.
I wanted to add a "read-chars" function and some way of telling the size
of a file, but it's not easy to get the file size in UTF-8 codepoints
as far as I can tell. As such, a "read-bytes" function will instead be
provided and the conversion to chars can be done later.