Add 'join' function
Feel like I already have a function like this but I can't remember it.
I'll remove this one later or make it an alias later on if that turns out
to be the case.

Also testing a new style of documentation comment that will become relevant with
the next commit.