Add examples/docgen.scm
This source file contains a Scheme program which reads another Scheme source
file and generates a markdown string. The idea is that this could be used to
create the documentation for the standard library. It requires writing some
markdown inline in the document for best results.

It is a slow program right now, presumably because of append being implemented
in Lisp code in an inefficient manner.