Add doc comments to lists.scm and generate docs/lists.md
Kind of a test run. I dislike how the parameters aren't automatically
added to the documentation.

I'm also not sure how wikis on Github work and wether it'd be better if
the documentation was stored there, but this is simple and allows me to
have a basic Scheme program to develop.