Add define-syntax and syntax-rules
This also resulted in fixing a bug in eqv? which would crash if you compared
a list containing a list with another list.

The implementation of the pattern matching and such is basically me reverse
engineering it from reading tutorials on how to use syntax-rules, so
the algorithm is probably not that great, but it works well enough to
define cond using it instead of having it be part of the interpreter, so
I traded one hairball for another.