Add vector type
What I wanted to get out of this was mostly a way to do append without cons,
since cons results in a shit ton of extra slice allocations. But, using
the new sappend implementation in lists.scm crashed on the docgen example for
some reason, but seems to work fine for simpler examples.