Fix printing of lambdas
Previously the interpreter would freeze when trying to print lambdas, so
now it just prints <closure>.

I'm guessing the reason it froze is because the default for printing pointers
in Go is not to print the address but instead to print &{whatever} which resulted
in infinite recursion when printing the environment.

This might change in the future since environment might become an accessible type
and in that case it'd probably have a custom String() function.