Add eval, interaction-environment and null-env.
Seems to work fine, even with e.g. (eval (list + 1 2)) which biwascheme
doesn't work with. The environments are WIP, I'd like a better way to do
different versions in the future and move syntax out of Eval so you can
pick and choose what default syntax is available in different
environments. For example, the go keyword probably shouldn't be in a
"scheme-report-environment".