syntax-rules work
Now works properly when it comes to matching lists followed by ellipses.
This allows the implementation of the let keyword.
As usual I doubt this is completely correct, and apparently in R6RS there
is a "syntax-case" that is a more general form of syntax-rules that I'll
have to implement later so this might all be for nothing.